TD-1166: replace id attributes with data-testid for countdown timer and join code
