+++ apps/dialog/e2e/tests/learning-scenarios/create-learning-scenario.test.ts
@@ -50,13 +50,20 @@ test.describe('create, share, chat, delete', () => {
     await page.getByRole('button', { name: 'Jetzt bereitstellen' }).click();
 
     await page.waitForURL('/learning-scenarios/**/share');
-    const code = await page.locator('#join-code').textContent();
+    const code = await page.getByTestId('join-code').textContent();
 
-    const countDown = page.locator('#countdown-timer');
+    const countDown = page.getByTestId('countdown-timer');
     await expect(countDown).toBeVisible();
 
-    const qrCode = page.locator('#qr-code');
+    const qrCode = page.getByTestId('#qr-code');
     await expect(qrCode).toBeVisible();
+
+    // verify countdown is also shown on the overview list
+    await page.goto('/learning-scenarios');
+    await page.waitForURL('/learning-scenarios**');
+    const card = page.getByTestId('entity-card').filter({ hasText: data.name }).first();
+    await expect(card).toBeVisible();
+    await expect(card.getByRole('timer')).toBeVisible();
 
     // join chat as teacher
     await page.goto('/logout');
@@ -93,7 +100,7 @@ test.describe('create, share, chat, delete', () => {
 
     // get code
     await page.waitForURL('/learning-scenarios/**/share');
-    const code = await page.locator('#join-code').textContent();
+    const code = await page.getByTestId('join-code').textContent();
 
     // join chat as student
     await page.goto('/logout');
