+++ apps/dialog/src/app/(authed)/(dialog)/characters/editor/[characterId]/share/page.tsx
@@ -46,7 +46,6 @@ export default async function Page(props: PageProps<'/characters/editor/[charact
       <div className="mx-auto mt-2 flex flex-col justify-center items-center text-center w-full">
         <h1 className="text-4xl sm:text-5xl font-medium mb-10">{t('join')}</h1>
         <CountDownTimer
-          id="countdown-timer"
           leftTime={Math.max(leftTime, 0)}
           totalTime={(character.maxUsageTimeLimit ?? 0) * 60}
           stopWatchClassName="w-4 h-4"
@@ -64,7 +63,7 @@ export default async function Page(props: PageProps<'/characters/editor/[charact
             <div className="flex flex-col items-center gap-4">
               <p className="text-2xl sm:text-3xl">{t('enter-code')}</p>
               <div className="flex items-center gap-2">
-                <p id="join-code" className="text-3xl sm:text-5xl text-primary font-bold">
+                <p data-testid="join-code" className="text-3xl sm:text-5xl text-primary font-bold">
                   {formattedInviteCode}
                 </p>
                 <TelliClipboardButton
@@ -82,7 +81,11 @@ export default async function Page(props: PageProps<'/characters/editor/[charact
           <div className="hidden sm:block w-1 border-r" />
           <section className="flex flex-col justify-between items-center gap-8 sm:gap-12">
             <h2 className="text-2xl sm:text-3xl text-center">{t('use-qr')}</h2>
-            <QRCodeSVG id="qr-code" className="w-64 h-64 sm:w-100 sm:h-100" value={shareUrl} />
+            <QRCodeSVG
+              data-testid="qr-code"
+              className="w-64 h-64 sm:w-100 sm:h-100"
+              value={shareUrl}
+            />
           </section>
         </main>
       </div>
